BREAK BC: New plugin system for symfony.
There are major changes in this changeset.

All plugins are now hosted in the plugins/ directory under the project root dir. They are self-contained
Plugins are now local only. It's not possible to install a plugin globally (because of problems with the multiplication of directories to search, problems with write access to the PEAR directory, ...).
Old plugins are not compatible, so I bumped the symfony branch to 0.9.0.
You must change the include_path() in your config.php: change sf_lib_dir to sf_root_dir:

    set_include_path(
      sfConfig::get('sf_root_dir').PATH_SEPARATOR.
      sfConfig::get('sf_app_lib_dir').PATH_SEPARATOR.
      sfConfig::get('sf_symfony_lib_dir').DIRECTORY_SEPARATOR.'vendor'.PATH_SEPARATOR.
      get_include_path()
    );

You must change your propel.ini configuration file:

    propel.targetPackage       = lib.model

    propel.php.dir                 = ${propel.output.dir}

The plugin-install, plugin-uninstall and plugin-upgrade tasks does not take local or global anymore as their first parameter:

    symfony plugin-install symfony/sfTestPlugin

If you upgrade now, upgrade carefully!



git-svn-id: file:///Users/saruman/Temp/symfony_svn/trunk@1931 4396de85-42e0-4213-a939-0f47b466bcdf